Trees
Added functionality for the program to label trees. This came in the form of adding an attribute to the Data to return true or false, dependent on whether or not the desired graph is a tree. This then tells the program whether or not the value is removed from the labeling set (since trees have one more vertex than edge it is allowed to repeat one vertex, and since we split the workload up by the labeling set and give one element of it to each thread, it falls from that implementation that we're choosing one member of the labeling set to repeat)